tests/requirements.txt
pytest==8.4.2
pytest-cov==5.0.0
httpx==0.27.2
requests-mock==1.12.1
fastapi==0.115.5
setuptools==75.6.0
slowapi==0.1.9
